# Backend/src/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Cryptocurrency Portfolio Tracker",
    "description": "API for cryptocurrency portfolio management",
    "version": "1.0.0"
  },
  "basePath": "/",
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
    "/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Register a new user",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/RegisterRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "User registered successfully",
            "examples": {
              "application/json": {
                "message": "User registered successfully"
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "examples": {
              "application/json": {
                "error": "Invalid input"
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "examples": {
              "application/json": {
                "error": "Internal server error"
              }
            }
          }
        }
      }
    },
    "/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login a user",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/LoginRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Login successful",
            "schema": {
              "type": "object",
              "properties": {
                "access_token": {
                  "type": "string"
                }
              }
            },
            "examples": {
              "application/json": {
                "access_token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "examples": {
              "application/json": {
                "error": "Invalid credentials"
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "examples": {
              "application/json": {
                "error": "Internal server error"
              }
            }
          }
        }
      }
    },
    "/portfolio": {
      "get": {
        "tags": ["Portfolio"],
        "summary": "Get portfolio value",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Portfolio value retrieved successfully",
            "examples": {
              "application/json": {
                "portfolio_value": "15000 $"
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "examples": {
              "application/json": {
                "error": "Internal server error"
              }
            }
          }
        }
      }
    },
    "/portfolio/advanced": {
      "get": {
        "tags": ["Portfolio"],
        "summary": "Get advanced portfolio value",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Advanced portfolio value retrieved successfully",
            "examples": {
              "application/json": {
                "portfolio_value": "15000 $",
                "details": {
                  "bitcoin": {
                    "amount": 1.5,
                    "value": "7500 $"
                  },
                  "ethereum": {
                    "amount": 10,
                    "value": "5000 $"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "examples": {
              "application/json": {
                "error": "Internal server error"
              }
            }
          }
        }
      }
    },
    "/prices": {
      "get": {
        "tags": ["Transactions"],
        "summary": "View cryptocurrency prices",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Prices retrieved successfully",
            "examples": {
              "application/json": {
                "bitcoin": "34000 $",
                "ethereum": "2000 $",
                "litecoin": "150 $"
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "examples": {
              "application/json": {
                "error": "Internal server error"
              }
            }
          }
        }
      }
    },
    "/price/{crypto}": {
      "get": {
        "tags": ["Transactions"],
        "summary": "View the price of a specific cryptocurrency",
        "parameters": [
          {
            "name": "crypto",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The cryptocurrency to view the price of"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "The current price of the specified cryptocurrency in USD",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                },
                "prices": {
                  "type": "string"
                }
              }
            },
            "examples": {
              "application/json": {
                "message": "The current price of bitcoin in USD is : ",
                "prices": "34000 USD"
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "examples": {
              "application/json": {
                "error": "Internal server error"
              }
            }
          }
        }
      }
    },
    "/transactions": {
      "post": {
        "tags": ["Transactions"],
        "summary": "Add a new transaction",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/TransactionRequest"
            }
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Transaction added successfully",
            "examples": {
              "application/json": {
                "message": "Transaction added successfully"
              }
            }
          },
          "403": {
            "description": "Unauthorized",
            "examples": {
              "application/json": {
                "error": "Unauthorized"
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "examples": {
              "application/json": {
                "error": "Internal server error"
              }
            }
          }
        }
      },
      "get": {
        "tags": ["Transactions"],
        "summary": "View all transactions",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Transactions retrieved successfully",
            "examples": {
              "application/json": [
                {
                  "id": 1,
                  "user_id": 1,
                  "cryptocurrency": "bitcoin",
                  "amount": 0.5,
                  "transaction_type": "buy"
                },
                {
                  "id": 2,
                  "user_id": 1,
                  "cryptocurrency": "ethereum",
                  "amount": 5,
                  "transaction_type": "buy"
                }
              ]
            }
          },
          "500": {
            "description": "Internal server error",
            "examples": {
              "application/json": {
                "error": "Internal server error"
              }
            }
          }
        }
      }
    },
    "/transactions/update/{id}": {
      "put": {
        "tags": ["Transactions"],
        "summary": "Update a transaction",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/TransactionUpdateRequest"
            }
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Transaction updated successfully",
            "examples": {
              "application/json": {
                "message": "Transaction updated successfully"
              }
            }
          },
          "403": {
            "description": "Unauthorized",
            "examples": {
              "application/json": {
                "error": "Unauthorized"
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "examples": {
              "application/json": {
                "error": "Internal server error"
              }
            }
          }
        }
      }
    },
    "/transactions/delete/{id}": {
      "delete": {
        "tags": ["Transactions"],
        "summary": "Delete a transaction",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Transaction deleted successfully",
            "examples": {
              "application/json": {
                "message": "Transaction deleted successfully"
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "examples": {
              "application/json": {
                "error": "Internal server error"
              }
            }
          }
        }
      }
    },
    "/transactions/delete/all": {
      "delete": {
        "tags": ["Transactions"],
        "summary": "Delete all transactions",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "All transactions deleted successfully",
            "examples": {
              "application/json": {
                "message": "All transactions deleted successfully"
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "examples": {
              "application/json": {
                "error": "Internal server error"
              }
            }
          }
        }
      }
    },
    "/session_data": {
      "get": {
        "tags": ["Session"],
        "summary": "View session data",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Session data retrieved successfully",
            "examples": {
              "application/json": {
                "last_transaction": {
                  "cryptocurrency": "bitcoin",
                  "amount": 0.5,
                  "transaction_type": "buy"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "examples": {
              "application/json": {
                "error": "Internal server error"
              }
            }
          }
        }
      }
    },
    "/transactions/history": {
      "get": {
        "tags": ["Transactions"],
        "summary": "View transaction history",
        "parameters": [
          {
            "name": "start_date",
            "in": "query",
            "type": "string",
            "format": "date",
            "required": false
          },
          {
            "name": "end_date",
            "in": "query",
            "type": "string",
            "format": "date",
            "required": false
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Transaction history retrieved successfully",
            "examples": {
              "application/json": {
                "total_value": "15000 $",
                "total_gain_loss": "2000 $",
                "crypto_holdings": {
                  "bitcoin": 0.5,
                  "ethereum": 5
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "examples": {
              "application/json": {
                "error": "Internal server error"
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
    "RegisterRequest": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      },
      "required": ["username", "password"]
    },
    "LoginRequest": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      },
      "required": ["username", "password"]
    },
    "TransactionRequest": {
      "type": "object",
      "properties": {
        "user_id": {
          "type": "integer"
        },
        "cryptocurrency": {
          "type": "string"
        },
        "amount": {
          "type": "number"
        },
        "transaction_type": {
          "type": "string",
          "enum": ["buy", "sell"]
        }
      },
      "required": ["user_id", "cryptocurrency", "amount", "transaction_type"]
    },
    "TransactionUpdateRequest": {
      "type": "object",
      "properties": {
        "amount": {
          "type": "number"
        },
        "transaction_price": {
          "type": "number"
        }
      }
    }
  },
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Enter your JWT token in the format **Bearer <token>**"
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ]
}